+++ packages/core/demo/data/scatter.ts
@@ -9,7 +9,7 @@ export const doubleLinearScatterData = [
 	{ group: "Dataset 2", employees: 2000, sales: 34100 },
 	{ group: "Dataset 2", employees: 4000, sales: 23100 },
 	{ group: "Dataset 2", employees: 7000, sales: 14100 },
-	{ group: "Dataset 2", employees: 6000, sales: 53100 },
+	{ group: "Dataset 2", employees: 6000, sales: 53100 }
 ];
 
 export const doubleLinearScatterOptions = {
@@ -18,14 +18,14 @@ export const doubleLinearScatterOptions = {
 		bottom: {
 			title: "No. of employees",
 			mapsTo: "employees",
-			scaleType: "linear",
+			scaleType: "linear"
 		},
 		left: {
 			title: "Annual sales",
 			mapsTo: "sales",
-			scaleType: "linear",
-		},
-	},
+			scaleType: "linear"
+		}
+	}
 };
 
 export const scatterDiscreteData = [
@@ -48,7 +48,7 @@ export const scatterDiscreteData = [
 	{ group: "Dataset 4", key: "More", value: 1200 },
 	{ group: "Dataset 4", key: "Sold", value: 9000 },
 	{ group: "Dataset 4", key: "Restocking", value: 24000 },
-	{ group: "Dataset 4", key: "Misc", value: 3000 },
+	{ group: "Dataset 4", key: "Misc", value: 3000 }
 ];
 
 export const scatterDiscreteOptions = {
@@ -57,35 +57,51 @@ export const scatterDiscreteOptions = {
 		bottom: {
 			title: "2019 Annual Sales Figures",
 			scaleType: "labels",
-			mapsTo: "key",
+			mapsTo: "key"
 		},
 		left: {
-			mapsTo: "value",
-		},
-	},
+			mapsTo: "value"
+		}
+	}
 };
 
 export const scatterLongLabelDiscreteData = [
 	{ group: "Dataset 1", key: "Qty", value: 34200 },
-	{ group: "Dataset 1", key: "6591DA8668C339B1B39297C61091E320C35391AB7AFC15B469F96B8A2DD0C231", value: 23500 },
+	{
+		group: "Dataset 1",
+		key: "6591DA8668C339B1B39297C61091E320C35391AB7AFC15B469F96B8A2DD0C231",
+		value: 23500
+	},
 	{ group: "Dataset 1", key: "Sold", value: 53100 },
 	{ group: "Dataset 1", key: "Restocking", value: 42300 },
 	{ group: "Dataset 1", key: "Misc", value: 12300 },
 	{ group: "Dataset 2", key: "Qty", value: 34200 },
-	{ group: "Dataset 2", key: "6591DA8668C339B1B39297C61091E320C35391AB7AFC15B469F96B8A2DD0C231", value: 53200 },
+	{
+		group: "Dataset 2",
+		key: "6591DA8668C339B1B39297C61091E320C35391AB7AFC15B469F96B8A2DD0C231",
+		value: 53200
+	},
 	{ group: "Dataset 2", key: "Sold", value: 42300 },
 	{ group: "Dataset 2", key: "Restocking", value: 21400 },
 	{ group: "Dataset 2", key: "Misc", value: 0 },
 	{ group: "Dataset 3", key: "Qty", value: 41200 },
-	{ group: "Dataset 3", key: "6591DA8668C339B1B39297C61091E320C35391AB7AFC15B469F96B8A2DD0C231", value: 18400 },
+	{
+		group: "Dataset 3",
+		key: "6591DA8668C339B1B39297C61091E320C35391AB7AFC15B469F96B8A2DD0C231",
+		value: 18400
+	},
 	{ group: "Dataset 3", key: "Sold", value: 34210 },
 	{ group: "Dataset 3", key: "Restocking", value: 1400 },
 	{ group: "Dataset 3", key: "Misc", value: 42100 },
 	{ group: "LongLabelShouldBeTruncated", key: "Qty", value: 22000 },
-	{ group: "LongLabelShouldBeTruncated", key: "6591DA8668C339B1B39297C61091E320C35391AB7AFC15B469F96B8A2DD0C231", value: 1200 },
+	{
+		group: "LongLabelShouldBeTruncated",
+		key: "6591DA8668C339B1B39297C61091E320C35391AB7AFC15B469F96B8A2DD0C231",
+		value: 1200
+	},
 	{ group: "LongLabelShouldBeTruncated", key: "Sold", value: 9000 },
 	{ group: "LongLabelShouldBeTruncated", key: "Restocking", value: 24000 },
-	{ group: "LongLabelShouldBeTruncated", key: "Misc", value: 3000 },
+	{ group: "LongLabelShouldBeTruncated", key: "Misc", value: 3000 }
 ];
 
 export const scatterLongLabelDiscreteOptions = {
@@ -94,12 +110,12 @@ export const scatterLongLabelDiscreteOptions = {
 		bottom: {
 			title: "2019 Annual Sales Figures",
 			scaleType: "labels",
-			mapsTo: "key",
+			mapsTo: "key"
 		},
 		left: {
-			mapsTo: "value",
-		},
-	},
+			mapsTo: "value"
+		}
+	}
 };
 
 export const scatterTimeSeriesData = lineTimeSeriesData;
@@ -110,12 +126,12 @@ export const scatterTimeSeriesOptions = {
 		bottom: {
 			title: "2019 Annual Sales Figures",
 			scaleType: "time",
-			mapsTo: "date",
+			mapsTo: "date"
 		},
 		left: {
-			mapsTo: "value",
-		},
-	},
+			mapsTo: "value"
+		}
+	}
 };
 
 // scatter - empty state
@@ -126,12 +142,12 @@ export const scatterEmptyStateOptions = {
 		bottom: {
 			title: "2019 Annual Sales Figures",
 			scaleType: "time",
-			mapsTo: "date",
+			mapsTo: "date"
 		},
 		left: {
-			mapsTo: "value",
-		},
-	},
+			mapsTo: "value"
+		}
+	}
 };
 
 // scatter - skeleton
@@ -142,13 +158,13 @@ export const scatterSkeletonOptions = {
 		bottom: {
 			title: "2019 Annual Sales Figures",
 			scaleType: "time",
-			mapsTo: "date",
+			mapsTo: "date"
 		},
 		left: {
-			mapsTo: "value",
-		},
+			mapsTo: "value"
+		}
 	},
 	data: {
-		loading: true,
-	},
+		loading: true
+	}
 };