+++ packages/core/src/charts/bar-simple.ts
@@ -16,7 +16,7 @@ import {
 	Tooltip,
 	Legend,
 	LayoutComponent,
-	Skeleton,
+	Skeleton
 } from "../components/index";
 
 export class SimpleBarChart extends AxisChart {
@@ -44,8 +44,8 @@ export class SimpleBarChart extends AxisChart {
 			new SimpleBar(this.model, this.services),
 			new ZeroLine(this.model, this.services),
 			new Skeleton(this.model, this.services, {
-				skeleton: Skeletons.VERT_OR_HORIZ,
-			}),
+				skeleton: Skeletons.VERT_OR_HORIZ
+			})
 		];
 
 		const components: any[] = this.getAxisChartComponents(