+++ packages/core/src/tests/test-environment.ts
@@ -10,6 +10,9 @@ import { groupedBarData, groupedBarOptions } from "../../demo/data";
 export const data = groupedBarData as ChartData;
 export const options = Object.assign(groupedBarOptions, {
 	title: "My chart",
+	data: {
+		selectedGroups: ["Dataset 1", "Dataset 3"]
+	}
 }) as any;
 
 export class TestEnvironment {
@@ -22,7 +25,7 @@ export class TestEnvironment {
 
 		this.chart = new Charts.ScatterChart(holder, {
 			data: this.chartData,
-			options: this.chartOptions,
+			options: this.chartOptions
 		});
 	}